fix(login): logger remove sensitive info from data
