+++ src/login/bklogin/api/constants.py
@@ -11,8 +11,6 @@ specific language governing permissions and limitations under the License.
 """
 
 
-from __future__ import unicode_literals
-
 from bklogin.common.constants import enum
 
 ApiErrorCodeEnum = enum(
