+++ src/login/bklogin/api/utils.py
@@ -11,8 +11,6 @@ specific language governing permissions and limitations under the License.
 """
 
 
-from __future__ import unicode_literals
-
 from django.conf import settings
 from django.http import JsonResponse
