+++ src/login/bklogin/bk_i18n/__init__.py
@@ -10,6 +10,5 @@ an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express o
 specific language governing permissions and limitations under the License.
 """
 
-from __future__ import unicode_literals
 
 default_app_config = "bklogin.bk_i18n.apps.BkI18nAppConfig"
