+++ src/login/bklogin/bk_i18n/middlewares.py
@@ -11,8 +11,6 @@ specific language governing permissions and limitations under the License.
 """
 
 
-from __future__ import unicode_literals
-
 import pytz
 from django.conf import settings
 from django.utils import timezone, translation
