+++ src/login/bklogin/bk_i18n/signal_receivers.py
@@ -10,7 +10,6 @@ an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express o
 specific language governing permissions and limitations under the License.
 """
 
-from __future__ import unicode_literals
 
 from django.conf import settings
 from django.contrib.auth.signals import user_logged_in
