+++ src/login/bklogin/components/http.py
@@ -10,7 +10,6 @@ an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express o
 specific language governing permissions and limitations under the License.
 """
 
-from __future__ import unicode_literals
 
 import time
 from urllib.parse import urlparse
@@ -18,6 +17,7 @@ from urllib.parse import urlparse
 import requests
 from django.conf import settings
 
+from .util import _remove_sensitive_info
 from bklogin.common.log import logger
 
 # NOTE: all new components should use the http.py here!
@@ -67,7 +67,13 @@ def _http_request(method, url, headers=None, data=None, timeout=None, verify=Fal
         else:
             return False, {"error": "method not supported"}
     except requests.exceptions.RequestException as e:
-        logger.exception("http request error! %s %s, data: %s, request_id: %s", method, url, data, request_id)
+        logger.exception(
+            "http request error! %s %s, data: %s, request_id: %s",
+            method,
+            url,
+            _remove_sensitive_info(data),
+            request_id,
+        )
         return False, {"error": str(e)}
     else:
         # record for /metrics
@@ -82,7 +88,7 @@ def _http_request(method, url, headers=None, data=None, timeout=None, verify=Fal
             error_msg = (
                 "http request fail! %s %s, data: %s, request_id: %s, response.status_code: %s, response.body: %s"
             )
-            logger.error(error_msg, method, url, str(data), request_id, resp.status_code, content)
+            logger.error(error_msg, method, url, _remove_sensitive_info(data), request_id, resp.status_code, content)
 
             return False, {
                 "error": (
